% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acars.R
\docType{data}
\name{acarsGoogle}
\alias{acarsGoogle}
\title{Autonomous Vehicle Patent Data from Google Patents}
\format{A data frame with 316 observations on 9 variables.
\describe{
\item{\code{docNum}}{A published document number including the kind code, publication number,
and kind code for the patent document.}
\item{\code{title}}{The title of the invention.}
\item{\code{assignee}}{The owner of the document.}
\item{\code{inventors}}{The name(s) of the inventor(s), separated by commas.}
\item{\code{priorityDate}}{The earliest priority date on the application.}
\item{\code{dateFiled}}{Date the document was filed. They calll it filing/creation date.}
\item{\code{pubDate}}{Date document became publicly available.}
\item{\code{grantDate}}{Date the application became a grant. NA if there is no associated grant.}
\item{\code{googleURL}}{The link to the Google Patents page for the document.}
}}
\usage{
acarsGoogle
}
\description{
An example data set of autonomous vehicle IP from major assignees.
}
\details{
The first row in the raw CSV export contains the search URL and is skipped.

The data search was performd on Saturday, March 18, 2017 from patents.google.com, and the exact
search: \href{https://patents.google.com/?q=AB\%3dautonomous&assignee=Apple*,Google*,Waymo*,Tesla*,Ford*,General*&before=filing:20170318}{Google Patents Search}
For all countries available on Google.

You process the raw data with the following commands:

\code{temp <- system.file("extdata", "google_autonomous_search.csv", package = "patentr")}

\code{# from the source package you can navigate to }

\code{temp <- read.csv("inst/extdata/google_autonomous_search.csv", skip = 1, stringsAsFactors = FALSE)}

\code{names(temp) <- googleNames}

\code{temp <- data.frame(lapply(temp, function(x){iconv(x,to="ASCII")}),stringsAsFactors = FALSE)}

\code{dateFields <- c("priorityDate","dateFiled","pubDate","grantDate")}

\code{temp[dateFields] <- as.data.frame(lapply(temp[dateFields], as.Date, format="\%m/\%d/\%y"))}
}
\seealso{
\url{https://patents.google.com/} 

\code{\link{acars}} provides a similar search from Sumobrain. 
\code{\link{acarsLens}} provides a simialr search from Lens.org.
}
\keyword{data}
