% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acars.R
\docType{data}
\name{acarsLens}
\alias{acarsLens}
\title{Autonomous Vehicle Patent Data from Lens Patent Search}
\format{A data frame with 863 observations on 26 variables.
\describe{
\item{resultNum}{The search result number.}
\item{countryCode}{The jurisdiction of the patent document.}
\item{\code{kindCode}}{The kind code.}
\item{docNum}{The published document number with country code and kind code included.}
\item{lensID}{The unique identification number of the document on lens.org}
\item{pubDate}{Date the document was published.}
\item{pubYear}{Year the document published.}
\item{appNum}{The filing number of the application (country code, number, and abridged kind code, typically 'A')}
\item{dateFiled}{Date the application for the patent document was filed.}
\item{priorityApps}{Applications this patent document claims priority. 
Format: Country code, application number, A = application or P = provisional, YYYYMMDD of priority. 
Multiple application separated by a double semi-colon.}
\item{title}{The title of the document.}
\item{assignee}{The name of the applicant(s) at the time of filing.}
\item{inventors}{The inventor(s).}
\item{lensURL}{The lens.org URL for the document.}
\item{docTypeLens}{A lens.org mapping of the doc type. 
Granted, application, ambiguous, unknown, search report, and possibly more values.}
\item{hasFullText}{A logical value to show if there is a full text available from lens.org}
\item{citeCount}{The number of times this document is cited, also known as forward citations.}
\item{familySimpleCount}{The number of unique documents in the immediate patent family.}
\item{familyExtendedCount}{The number of unique documents sharing a priority applicaiton in the extended family.}
\item{seqCount}{Used in biological applications -- the number of sequences in the application.}
\item{cpcClasses}{The CPC classification codes, separated by a double semi-colon.}
\item{ipcrClasses}{The IPCR classification codes, separated by a double semi-colon.}
\item{usClasses}{The US classification codes, separated by a double semi-colon.}
\item{pubmedID}{A pubmed ID to any related research.}
\item{DOI}{A digital object identifier. 
Go to doi.org and paste the value to get the associated research paper.}
\item{npl}{Non-patent literature, or citations of non-patent sources.
Separated with double semi-colons.}


}}
\usage{
acarsLens
}
\description{
An example data set of autonomous vehicle IP from major assignees.
}
\details{
The data search was performd on Saturday, March 18, 2017 from lens.org, and the exact
search:

\href{https://www.lens.org/lens/search?q=abstract\%3Aautonomous+\%26\%26+applicant\%3A\%28Apple*+OR+Google*+OR+Waymo*+OR+Tesla*+OR+Ford*+OR+General*\%29&predicate=\%26\%26&l=en}{Lens Patents Search}

For all countries available on Lens.

Can get raw data with the following commands:

\code{temp <- system.file("extdata", "lens_autonomous_search.csv", package = "patentr")}

\code{temp <- read.csv(temp, stringsAsFactors = FALSE)}

\code{temp <- data.frame(lapply(temp, function(x){iconv(x,to="ASCII")}),stringsAsFactors = FALSE)}
 
\code{names(temp) <- lensNames}

\code{temp$dateFiled <- as.Date(temp$dateFiled, format = '\%m/\%d/\%y')}

\code{temp$pubDate <- as.Date(temp$pubDate, format='\%m/\%d/\%y')} # note that % y is system-specific and may not work everywhere.

\code{colsNum <- c("resultNum","citeCount","familySimpleCount","familyExtendedCount", "seqCount")}

\code{temp[colsNum] <- sapply(temp[colsNum], as.numeric)}

\code{temp$hasFullText <- sapply(temp$hasFullText, function(x) ifelse(x=="yes",TRUE,FALSE))}
}
\seealso{
\url{www.lens.org} You can export without an account, or can create 
an account to save your searches. 

\code{\link{acarsGoogle}} provides a similar search from Google. 
\code{\link{acars}} provides a similar search from sumobrain.
}
\keyword{data}
