% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPatentData.R
\name{addPdfImage}
\alias{addPdfImage}
\title{Make a PDF output of a plot}
\usage{
addPdfImage(graph, name = "image", w = 12, h = 12)
}
\arguments{
\item{graph}{The graph object to input}

\item{name}{A character name to name your file. It can have a filepath as well.}

\item{w}{The width, in inches, of your image, default set to 12.}

\item{h}{The height, in inches, of your image, default set to 12.}
}
\value{
No ret
}
\description{
Make a PDF output of a plot.
}
\examples{

sumo <- cleanPatentData(patentData = patentr::acars, columnsExpected = sumobrainColumns,
cleanNames = sumobrainNames,
dateFields = sumobrainDateFields,
dateOrders = sumobrainDateOrder,
deduplicate = TRUE,
cakcDict = patentr::cakcDict,
docLengthTypesDict = patentr::docLengthTypesDict,
keepType = "grant",
firstAssigneeOnly = TRUE,
assigneeSep = ";",
stopWords = patentr::assigneeStopWords)

# df <- dplyr::select(sumo, title, abstract)
df <- sumo[,c("title","abstract")]
addPdfImage(wordCloudIt(df, excludeWords, minfreq = 20,
                        random.order = FALSE, rot.per = 0.25),"wordCloud")

}
