% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assigneeStopWords}
\alias{assigneeStopWords}
\title{A simple stop word list for assignee names.}
\format{A character vector

\describe{
\item{assigneeStopWords}{A character vector of stop words.}
}}
\usage{
assigneeStopWords
}
\description{
A character vector of common stop words to remove from assignee names for 
name standardization, such as "inc".
}
\seealso{
\code{\link{cleanNames}}
}
\keyword{data}
