% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frequency.R
\name{check_frequency}
\alias{check_frequency}
\title{Check the frequency of a column.}
\usage{
check_frequency(
  dataset,
  column,
  decreasing = T,
  graph = TRUE,
  fill = "pink",
  num = 10
)
}
\arguments{
\item{dataset}{The dataset to use.}

\item{column}{The name of the column to look at.}

\item{decreasing}{A logical value to determine the order of the frequency table. Default to TRUE.}

\item{graph}{A logical value to determine whether to plot the frequency table. Default to TRUE.}

\item{fill}{The color to plot. Default to pink.}

\item{num}{The number of top items to plot. Default to 10.}
}
\value{
\code{dataset}: The frequency table of the column.
}
\description{
This function will check the frequency of a column in the dataset. By default, the function gives a frequency table in decreasing order.
}
\examples{
library(patents)
data('patent_data')
x1=clean_by(patent_data, "assigneeCity")
x1=check_frequency(x1, "assigneeCity")
}
