% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPatentData.R
\name{chooseFiles}
\alias{chooseFiles}
\title{Allow the user to navigate to files manually.}
\usage{
chooseFiles()
}
\value{
A list of character vectors with absolute pathnames to files.
}
\description{
Uses a popup window (Tk file dialog) to allow the user to choose a list of zero or more files interactively.
}
\examples{
\dontrun{
filePaths <- chooseFiles()
allData <- readIpExcel(filePaths)
}
}
