% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_by.R
\name{clean_by}
\alias{clean_by}
\title{Clean the dataset based on a column.}
\usage{
clean_by(dataset, column)
}
\arguments{
\item{dataset}{The dataset to clean.}

\item{column}{The name of the column to clean. It can be \code{inventorState} or \code{assigneeState}.}
}
\value{
\code{dataset}: The cleaned dataset. Usually have more rows than the original dataset.
}
\description{
This function will clean the dataset based on the columns about demographic data, including: \code{inventorState}, \code{assigneeState}, \code{inventorName}, \code{assigneeName}, \code{inventorCity}, and \code{assigneeCity}.
If the column is \code{inventorState} or \code{assigneeState}, it will separate the column into multiple rows and clean the duplicates and other countries.
}
\examples{
library(patents)
data('patent_data')
x1=clean_by(patent_data, "inventorState")
x2=clean_by(patent_data, "assigneeState")
x3=clean_by(patent_data, "inventorsName")
x4=clean_by(patent_data, "assigneeName")
x5=clean_by(patent_data, "inventorCity")
x6=clean_by(patent_data, "assigneeCity")
}
