% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPatentData.R
\name{cleanGoogleURL}
\alias{cleanGoogleURL}
\title{Sanitize a Google URL before attempting to extract data}
\usage{
cleanGoogleURL(googleURL, langCode = "en")
}
\arguments{
\item{googleURL}{A character value of a google URL.}

\item{langCode}{A language code, default set to "en" English.}
}
\value{
A clean character vector of a Google Patents URL.
}
\description{
Clean up the google URL to make sure it will be read properly.

If you use the \code{\link{createGoogleURL}} function, you won't have to use this function. 
However, if you use your own generator or want to change the language, use this 
function to do so.
}
\examples{

cleanGoogleURL("https://patents.google.com/patent/US8818682B1/mx")
cleanGoogleURL("https://patents.google.com/patent/US8818682B1/")
cleanGoogleURL("https://patents.google.com/patent/US8818682B1")
cleanGoogleURL("https://patents.google.com/patent/US8818682B1/en")

}
\seealso{
\code{\link{createGoogleURL}}
}
