% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{cleanNames}
\alias{cleanNames}
\title{Clean up string names.}
\usage{
cleanNames(rawNames, firstAssigneeOnly = TRUE, sep = ";",
  removeStopWords = TRUE, stopWords = patentr::assigneeStopWords)
}
\arguments{
\item{rawNames}{The character vector you want to clean up}

\item{firstAssigneeOnly}{A logical value, default set to TRUE, keeping only the first 
assignee if multiple exist.}

\item{sep}{The separating character for multiple assignees, default set to semi-colon.}

\item{removeStopWords}{Logical default TRUE, if want to remove common company stopwords 
found in the \code{stopWords} parameter.}

\item{stopWords}{An optional character vector of words you want to remove. Default to 
\code{\link{assigneeStopWords}}.}
}
\value{
A character vector of cleaned up character names.
}
\description{
Quick cleanup of characters in a string, 
typically assignee (company names) and the inventors. 

If you have issues with this, you may need to convert to UTF-8 or ASCII.
Use the \code{iconv(thisVector, to="UTF-8")} or \code{to="ASCII"} and it should 
fix the problem. See the examples for the code.


This function: 
\enumerate{
\item{Removes values between spaces, such as (US)}
\item{Changes all names to lower case}
}
}
\examples{

assigneeNames <- cleanNames(acars$assignee)
# get a feel for the less-messy data
head(sort(table(assigneeNames), decreasing = TRUE))

# for a messier example, note you need to convert to ASCII/UTF-8 to get rid of errors
# associated with tolower
rawGoogleData <- system.file("extdata", "google_autonomous_search.csv", package = "patentr")
rawGoogleData <- read.csv(rawGoogleData, stringsAsFactors = FALSE, skip = patentr::skipGoogle)
rawGoogleData <- data.frame(lapply(rawGoogleData, 
function(x){iconv(x, to = "ASCII")}), stringsAsFactors = FALSE)
assigneeClean <- cleanNames(rawGoogleData$assignee)
head(sort(table(assigneeClean), decreasing = TRUE))

}
