% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{cleanSumobrainNames}
\alias{cleanSumobrainNames}
\title{Generate a standard set of header names for Sumobrain import data}
\usage{
cleanSumobrainNames(sumobrainData = NA, columnsExpected = 11,
  cleanNames = c("docNum", "docType", "pubDate", "title", "abstract",
  "inventors", "assignee", "appNum", "dateFiled", "classPrimary",
  "classOthers"))
}
\arguments{
\item{sumobrainData}{A data frame 11 columns wide, imported from sumobrain.com.}

\item{columnsExpected}{A sumobrain export has 11 columns. In case their export changes,
you can modify the value.}

\item{cleanNames}{A standard list of clean names.}
}
\value{
A data frame 11 columns wide, with standard column names used in other
functions.
}
\description{
Create a standard nameset from Sumobrain import data. 
See \code{\link{acars}} for the name set.

To download Sumobrain data, go to \url{http://www.sumobrain.com} and create a free
account. Then run your search, export the data (250 max at a time), and use the 
\code{\link{chooseFiles}} and \code{\link{importPatentData}} functions to load
the data into R.
}
\examples{
cleanData <- cleanSumobrainNames(sumobrainData = acars)

}
