% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{createGoogleURL}
\alias{createGoogleURL}
\title{Create a URL link to Google patents.}
\usage{
createGoogleURL(countryCode, pubNum, kindCode,
  googleURL = "https://patents.google.com/patent/")
}
\arguments{
\item{countryCode}{A character vector of the country code of the document. 
Typically a two-letter character.}

\item{pubNum}{A character vector of the numeric portion of a publication number.}

\item{kindCode}{character vector of the kind code of a document. If not available,
enter a blank string "".}

\item{googleURL}{A character string of the URL to Google Patents, with working
default value.}
}
\value{
A character vector of properly formatted URL strings.
}
\description{
Create a URL string to link you to Google Patents. 

By concatenating the country code, publication number, and kind code, you can
generate a URL to link you to google patents for further exploration. This 
feature is especially useful when browsing the data in a spreadsheet or in 
a Shiny app. It is also useful for extracting data from the HTML content. 

As each website (Google, lens.org, sumobrain.com, etc..) has a different 
method for generating patent URLs, these functions are website-specific. 

The original Google patents version still works as of March 2017 and the 
\code{googleURL} value is  \code{https://www.google.com/patents/}. This older 
version may be easier to extract data.
}
\examples{
acars$kindCode <- extractKindCode(acars$docNum)
acars$pubName <- extractPubNumber(acars$docNum)
acars$googleURL <- createGoogleURL(countryCode = acars$countryCode, 
pubNum = acars$pubNum, kindCode =acars$kindCode)
head(acars$googleURL)

}
