% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docLengthTypes}
\alias{docLengthTypes}
\title{A document mapper for country codes and document digit length to the type of 
document.}
\format{A data frame with a key and value pair. 

\describe{
\item{key}{A concatenated country code and length of the numeric portion of a 
document number. For example: US7 is a US document with 7 digits.}
\item{value}{The type of patent document based on the country code and document 
length value.}

}}
\usage{
docLengthTypes
}
\description{
A simple table that helps map the country code and length of the numeric portion 
of the data to the type of document.
}
\keyword{data}
