% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docLengthTypesDict}
\alias{docLengthTypesDict}
\title{A named vector of key/value pairs for country codes and publication number 
document lengths used to determine the type of document.}
\format{A named character vector

\describe{
\item{docLengthTypesDict}{A named character vector representing key/value pairs
of doc lengths, country codes, and type of patent document.}

}

Built with the following code: 

\code{docLengthTypesDict <- docLengthTypes$value}

\code{names(docLengthTypesDict) <- docLengthTypes$key}}
\usage{
docLengthTypesDict
}
\description{
A named vector of key/value pairs for country codes and publication number 
document lengths used to determine the type of document.
}
\seealso{
\code{\link{generateDocType}}, \code{\link{docLengthTypes}}
}
\keyword{data}
