% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_by_date.R
\name{dummy_by_date}
\alias{dummy_by_date}
\title{Create a dummy variable by date.}
\usage{
dummy_by_date(data, date_column, cutoff_date, dummy = "dummy_code")
}
\arguments{
\item{data}{The dataset to use}

\item{date_column}{The date column to look at. Should be a string.}

\item{cutoff_date}{The cutoff date. Should be a string or a Date object.}

\item{dummy}{The name of the dummy variable to create. Should be a string.}
}
\value{
\code{dataset}: The dataset with the new dummy variable.
}
\description{
This function creates a dummy variable based on date. It will assign 1 to the dummy variable if the date is before the cutoff date.
}
\examples{
library(patents)
library(tidyverse)
data('patent_data')
x1 = dummy_by_date(patent_data, "datePublished", "2010-01-01", "dummy_date")
xx1 = x1[x1$dummy_date == 1,]
xx1 = x1[x1$dummy_date != 1,]

data = patent_data \%>\% arrange(datePublished) \%>\% filter(!is.na(applicationFilingDate))
x2 = dummy_by_date(data, "applicationFilingDate", "1900-01-01", "dummy_date")
xx2 = x2[x2$dummy_date == 1,]
xx2 = x2[x2$dummy_date != 1,]
}
