% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excludeWords}
\alias{excludeWords}
\title{A standard list of words to exclude in a patent word cloud.}
\format{A character vector. 

\describe{
\item{excludeWords}{A character vector of words to exclude}
}}
\usage{
excludeWords
}
\description{
A standard list of words to exclude from a patent data word cloud.
}
\keyword{data}
