% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{extractCleanDate}
\alias{extractCleanDate}
\title{Format patent dates.}
\usage{
extractCleanDate(dateVector, orders = "ymd")
}
\arguments{
\item{dateVector}{A vector of character dates.}

\item{orders}{The orders the dates appear in. 
Sumobrain is "ymd" and Lens.org and Google data are "mdy". Hardcoded values include 
\code{\link{googleDateOrder}},\code{\link{lensDateOrder}}, 
and \code{\link{sumobrainDateOrder}}.}
}
\value{
A date vector of year, month, day dates.
}
\description{
Create a clean year, month, day date.

Reading data in and aout of R may cause date mistakes, using a simple set
function will ensure data types are the right format and class type. This 
data format is cleaned up to be in the format yyyy-mm-dd with no hours,
minutes, seconds, or time zone attached.
}
\examples{
acars$pubDate <- extractCleanDate(dateVector = acars$pubDate, orders = "ymd")


}
