% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{extractCountryCode}
\alias{extractCountryCode}
\title{Extract the country code from a vector string of document numbers.}
\usage{
extractCountryCode(docNum)
}
\arguments{
\item{docNum}{The character vector of document numbers.}
}
\value{
A character vector of the same length inputted, with 2-4 characters
representing the country code of the ptaent document.
}
\description{
Extract the country code from a patent document number, which is the 
first two to four letters in a patent document number.

For example, if a patent number is US8880270, the country code is US. In rare
cases, we have more than two letters. Typical country codes are US (United States),
EP (Europe), JP (Japan), and WO (World, meaning a PCT application).
}
\examples{
# create a new column called countryCode in the acars data set
acars$countryCode <- extractCountryCode(acars$docNum)
head(acars[,c("docNum","countryCode")])

}
