% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{extractDocLength}
\alias{extractDocLength}
\title{Get a code for length of doc and country code}
\usage{
extractDocLength(countryCode, pubNum)
}
\arguments{
\item{countryCode}{A string vector of country codes}

\item{pubNum}{A string vector of the numeric portion of a publication number.}
}
\value{
A string vector of concatenated country code and publication number 
length, such as US11 or EP9.
}
\description{
Generate a custom concatenation of country code and length of 
the publication number, for document type identification purposes. 

Given limited metadata available on free sites, often times the downloaded
data set does not include the type of patent document. There are two easy ways to 
discover the type of a patent document. A dictionary stored with the 
package can compare the output to match up the type of patent document. 

\enumerate{
\item The kind code, if present, is typically the same for each country.
\code{B} is usually a patent and \code{A} is usually an application.
\item The length of the publication number, along with the country code, is 
another great indicator. Applications in USA have 11 numbers, and, for now,
9 numbers for granted patents.
}
}
\examples{
acars$pubNum <- extractPubNumber(acars$docNum)
acars$countryCode <- extractCountryCode(acars$docNum)
acars$officeDocLength <- extractDocLength(countryCode = acars$countryCode,
pubNum = acars$pubNum)
head(acars[,c("officeDocLength","docNum")])

}
