% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{extractKindCode}
\alias{extractKindCode}
\title{Extract the kind code, if available, from the publication number.}
\usage{
extractKindCode(docNum)
}
\arguments{
\item{docNum}{The character vector of document numbers.}
}
\value{
A character vector of kind codes. If none found, a blank character is returned.
}
\description{
Extracts the kind code, a one-to-two character code with a letter and 
typically a number, if found in the document (published) number.
}
\examples{
acars$kindCode <- extractKindCode(acars$docNum)
head(acars[,c("docNum","kindCode")]) 
  
}
