% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{extractPubNumber}
\alias{extractPubNumber}
\title{Extract the numeric portion of the document (published) number.}
\usage{
extractPubNumber(docNum)
}
\arguments{
\item{docNum}{The character vector of document numbers.}
}
\value{
A character vector of same length inputted, with varying length. 
Typical lengths are 7 to 11 characters. Only numbers are returned. All other
characters are stripped.
}
\description{
Extract the numeric portion of the document number. 
This is useful for a number of processing applications, and thus is beneficial
to isolate from the entire publication number.
}
\examples{
acars$pubNum <- extractPubNumber(acars$docNum)
head(acars[,c("docNum","pubNum")]) 
  
 
}
