% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePatentData.R
\name{factorForGraph}
\alias{factorForGraph}
\title{Factor one column by another column's popularity}
\usage{
factorForGraph(df, xVal, fillVal)
}
\arguments{
\item{df}{A data frame containing the x and fill value columns.}

\item{xVal}{A character value from a header name in \code{df} 
that will be used as 
the x value in a ggplot2 plot.}

\item{fillVal}{A character value from a header name in \code{df} 
that will be used as the 
fill value in a ggplot2 plot.}
}
\value{
A data frame with one factored column.
}
\description{
Factor (or refactor) a data frame of values to be used 
for graphing in the correct order.

Many graphs require a reordering when plotting with a fill value. This 
helper function factors the x-value of a plot that will be stacked by 
\code{fillVal}.
}
\examples{


sumo <- cleanPatentData(patentData = patentr::acars, columnsExpected = sumobrainColumns,
cleanNames = sumobrainNames,
dateFields = sumobrainDateFields,
dateOrders = sumobrainDateOrder,
deduplicate = TRUE,
cakcDict = patentr::cakcDict,
docLengthTypesDict = patentr::docLengthTypesDict,
keepType = "grant",
firstAssigneeOnly = TRUE, 
assigneeSep = ";",
stopWords = patentr::assigneeStopWords)

# note that in reality, you need a patent analyst to carefully score
# these patents, the score here is for demonstrational purposes
score <- round(rnorm(dim(sumo)[1],mean=1.4,sd=0.9))
score[score>3] <- 3
score[score<0] <- 0
sumo$score <- score
dim(sumo)
sumoFactor <- factorForGraph(sumo, "assigneeClean", "score")
# if you want to view, uncomment and load ggplot2
# ggplot(sumoFactor, aes(x=assigneeClean, y=score, fill=factor(score))) + 
# geom_bar(stat="identity")


}
