% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePatentData.R
\name{flippedHistogram}
\alias{flippedHistogram}
\title{Plot a flipped histogram with a fill value}
\usage{
flippedHistogram(df, xVal, fillVal, colors = patentr::scoreColors)
}
\arguments{
\item{df}{The original data frame of patent data}

\item{xVal}{A character value of a name in \code{df}}

\item{fillVal}{A character value of a name in \code{df} to color the chart.}

\item{colors}{A character vector of colors, the same length as the number of 
unique values in the column of \code{xVal[,fillVal]}. Default set to 
\code{scoreColors}}
}
\value{
A plot
}
\description{
Plot a flipped histogram with fill values.

Often times, you want to plot a histogram showing patent documents 
faceted by one value and filled by another.
}
\examples{


sumo <- cleanPatentData(patentData = patentr::acars, columnsExpected = sumobrainColumns,
cleanNames = sumobrainNames,
dateFields = sumobrainDateFields,
dateOrders = sumobrainDateOrder,
deduplicate = TRUE,
cakcDict = patentr::cakcDict,
docLengthTypesDict = patentr::docLengthTypesDict,
keepType = "grant",
firstAssigneeOnly = TRUE,
assigneeSep = ";",
stopWords = patentr::assigneeStopWords)

# note that in reality, you need a patent analyst to carefully score
# these patents, the score here is for demonstrational purposes
score <- round(rnorm(dim(sumo)[1],mean=1.4,sd=0.9))
score[score>3] <- 3
score[score<0] <- 0
sumo$score <- score
sumo$assigneeSmall <- strtrim(sumo$assigneeClean,12)
flippedHistogram(sumo, "assigneeSmall","score",colors=scoreColors)


}
