% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_by_group.R
\name{freq_by_group}
\alias{freq_by_group}
\title{Check the frequency of a column by groups.}
\usage{
freq_by_group(
  dataset,
  group,
  column,
  decreasing = TRUE,
  graph = TRUE,
  num = 10,
  fill = "pink"
)
}
\arguments{
\item{dataset}{The dataset to use.}

\item{group}{The name of the column to group by. Not a character input. Use the column name directly.}

\item{column}{The name of the column to look at.}

\item{decreasing}{A logical value to determine the order of the frequency table. Default to TRUE.}

\item{graph}{A logical value to determine whether to plot the frequency table. Default to TRUE.}

\item{num}{The number of top items to plot. Default to 10.}

\item{fill}{The color to plot. Default to pink.}
}
\value{
\code{freq_list}: A list of frequency tables for each group.
}
\description{
This function will check the frequency of a column by groups in the dataset.
By default, the function gives a frequency table in decreasing order.
The function also plots the frequency table by groups.
}
\examples{
library(patents)
data('patent_data')
x1=clean_by(patent_data, "inventorState")
x1=freq_by_group(x1, inventorState, "inventorsName", graph=FALSE)
}
