% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPatentData.R
\name{getClaimsText}
\alias{getClaimsText}
\title{Get claims data for all rows in a data frame}
\usage{
getClaimsText(googleURLs, langCode = "en", whichClaim = 1)
}
\arguments{
\item{googleURLs}{A character vector of Google URLs}

\item{langCode}{A language code, default set to "en"}

\item{whichClaim}{Which claim (if available) to return. Default set to 1st.}
}
\description{
Generate claims data for all rows in a data frame. 

This is a wrapper function for the \code{\link{getClaimFromURL}} function.
}
\examples{

cc <- extractCountryCode(acars$docNum)
pn <- extractPubNumber(acars$docNum)
kc <- extractKindCode(acars$docNum)
urls <- createGoogleURL(countryCode = cc, pubNum = pn ,kindCode = kc)
urls <- urls[1:4]
clms <- getClaimsText(urls)
clms[1]

}
\seealso{
\code{\link{createGoogleURL}}, \code{\link{cleanGoogleURL}},
\code{\link{getClaimFromURL}}
}
