% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{googleColumns}
\alias{googleColumns}
\title{Number of columns in Google Patents export data.}
\format{A numeric value.

\describe{
\item{googleColumns}{A numeric value of number of columns in a csv export from 
Google Patents.}

}}
\usage{
googleColumns
}
\description{
The number of columns in a Google Patents CSV export.
}
\details{
Used with \code{\link{acarsGoogle}} data.
}
\seealso{
\code{\link{skipGoogle}}, \code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}}, \code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
