% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{googleDateFields}
\alias{googleDateFields}
\title{A simple list of date column names in Google patent data.}
\format{A character vector.

\describe{
\item{googleDateFields}{A character vector of date fields.}
}}
\usage{
googleDateFields
}
\description{
A character vector of date fields in Google patent data.
}
\seealso{
\code{\link{cleanHeaderNames}}
}
\keyword{data}
