% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{googleDateOrder}
\alias{googleDateOrder}
\title{Date order for Google Patents data.}
\format{A character value.

\describe{
\item{googleDateOrder}{A character variable of date order.}
}}
\usage{
googleDateOrder
}
\description{
A date order to be used in Google patent date data.
}
\seealso{
\code{\link{extractCleanDate}}
}
\keyword{data}
