% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{googleNames}
\alias{googleNames}
\title{Header names for a data upload sourced from Google Patents data exports.}
\format{A character vector.

\describe{
\item{googleNames}{A 9-element character vector of clean Google patent names.}

}}
\usage{
googleNames
}
\description{
A character vector to standardize the headers of the imported csv from a 
Google Patents patent data export. Used with \code{\link{acarsGoogle}} data.
}
\seealso{
Go to \href{patents.google.com}{Google Patents} to download the data.

\code{\link{skipGoogle}}, \code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}},\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}}, \code{\link{lensNames}}
}
\keyword{data}
