% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPatentData.R
\name{importPatentData}
\alias{importPatentData}
\title{Read in a data file or list of files from excel spreadsheets.}
\usage{
importPatentData(rawDataFilePath = NA, skipLines = 1)
}
\arguments{
\item{rawDataFilePath}{A filepath, or list of filespaths, for xls files.}

\item{skipLines}{Number of lines to skip before reading in your data file.}
}
\value{
A single data frame of all data. NULL if no data.
}
\description{
Import, read, and connect patent data files. Currently: xls files
from a filepath. 
Future use: can read from a URL, an xlsx file, google doc, and a csv.
}
\examples{
\dontrun{

# access the files here and put them in a data/ folder of your working directory.
file1 <- system.file("extdata", "sumobrain_autonomous_search1.xls", package = "patentr")
file2 <- system.file("extdata", "sumobrain_autonomous_search2.xls", package = "patentr")

# assume csv files are in the data folder
ipData <- readPatentData(rawDataFilePath = list.files('data/', full.names=T), skipLines = 1)
}


}
