% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kindCodes}
\alias{kindCodes}
\title{A kind codes database to show the type of document for each patent document.}
\format{A data frame.

\describe{
\item{\code{countryCode}}{The country code for the originating office where the application 
was filed.}
\item{kindCode}{The letter/number code to signify the type of document. Codes may 
change after a certain date, so pay attention to \code{dateStarted} and \code{dateDeprecated}}
\item{isDeprecated}{Logical TRUE/FALSE if the kind code for the country is no longer in use.}
\item{dateDeprecated}{The date the kind code stopped being in use.}
\item{isNew}{If the kind code is a replacement for a former kind code, TRUE, else FALSE.}
\item{dateStarted}{If isNew == TRUE, the date the new kind code began being used.}
\item{comment}{Additional information explaining the details of the kind code.}
\item{docTypeLong}{The long version of the document type.}
\item{docType}{A shorter, standardized version of \code{docTypeLong}.}
\item{expectDuplicate}{A logical TRUE/FALSE to help the analyst understand if the 
published document is expected to have a duplicate publication. For example, USB2 is 
a granted patent that has an application that was also published, whereas USB1 has no 
previous documents published. This helps speed up the deduplication process. }
\item{countryAndKindCode}{A concatenation of country code and kind code. Used in 
the deduplication process and to determine the type of document.}
}}
\usage{
kindCodes
}
\description{
Patent documents have associated kind codes, which are letter/number code 
combinations that signify the type of document, such as application, granted 
patent, utility patent, etc. These kind codes vary by country and are a useful 
approach to classifying patent document types. Most, however, not all, downloaded 
data from free services such as sumobrain.com or lens.org includes the kind code 
at the end of the patent document number.
}
\details{
For example, from the sumobrain.com download from the \code{\link{acars}} data set,
here are three documents:
\enumerate{
\item{US6523912}
\item{US20030060197}
\item{EP1310400A1}
}

The first two items are missing kind codes. The third item has kind code A1 
and the country code is EP.

To clean the data yourself:

\code{temp <- readxl::read_excel(system.file("extdata", "kindCodes.xlsx", package = "patentr"))}

\code{temp <- replace(temp, is.na(temp), "NA")}

\code{temp$dateDeprecated <- as.numeric(temp$dateDeprecated)}

\code{temp$dateDeprecated <- as.Date(temp$dateDeprecated, origin = "1899-12-30")}

\code{temp$dateStarted <- as.numeric(temp$dateStarted)}

\code{temp$dateStarted <- as.Date(temp$dateStarted, origin = "1899-12-30")}

\code{temp$countryAndKindCode <- with(temp,paste0(countryCode, kindCode))}

See \url{https://www.r-bloggers.com/date-formats-in-r/} for excel mac/windows 
and confirm this origin works for you by reviewing the source file.

View the data sources:
\enumerate{
\item{\href{https://www.uspto.gov/learning-and-resources/support-centers/electronic-business-center/kind-codes-included-uspto-patent}{USPTO kind codes}}
\item{\href{https://worldwide.espacenet.com/help?locale=en_EP&method=handleHelpTopic&topic=kindcodes\%5C}{EPO file histories}}
\item{\href{https://www.cas.org/content/references/patkind}{CAS list of kind codes}}
\item{\href{http://ipbookcompanion.org/links/pk_codes.pdf}{IP Book kind codes}}
\item{\href{http://www.thomsonfilehistories.com/docs/RESOURCES_Kind\%20Codes\%20by\%20Country.pdf}{Thomson File Histories}}
}
}
\keyword{data}
