% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lensColumns}
\alias{lensColumns}
\title{The number of columns in a lens.org csv data export.}
\format{A numeric value.

\describe{
\item{lensColumns}{A numeric value of the number of columns in a lens.org 
patent data export. }

}}
\usage{
lensColumns
}
\description{
The number of columns in a lens.org csv data export.
}
\details{
Used with \code{\link{acarsLens}} data.
}
\seealso{
\code{\link{skipGoogle}}, \code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}}, \code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
