% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lensDateFields}
\alias{lensDateFields}
\title{A simple list of date column names in lens.org data.}
\format{A character vector.

\describe{
\item{lensDateFields}{A character vector of date fields.}
}}
\usage{
lensDateFields
}
\description{
A character vector of date fields in lens.org data.
}
\seealso{
\code{\link{cleanHeaderNames}}
}
\keyword{data}
