% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lensDateOrder}
\alias{lensDateOrder}
\title{Date order for lens.org data.}
\format{A character value.

\describe{
\item{lensDateOrder}{A character variable of date order.}
}}
\usage{
lensDateOrder
}
\description{
A date order to be used in lens.org date data.
}
\seealso{
\code{\link{extractCleanDate}}
}
\keyword{data}
