% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lensNames}
\alias{lensNames}
\title{Header names for a data upload sourced from lens.org.}
\format{A character vector.

\describe{
\item{sumobrainNames}{A 26-element character vector of clean lens.org header names.}

}}
\usage{
lensNames
}
\description{
A character vector to standardize the headers of the imported csv from a 
lens.org patent data export.
}
\details{
Used with \code{\link{acarsLens}} data.
}
\seealso{
Go to \href{lens.org}{Lens}, optionally create a free account, and 
download the data. 

\code{\link{skipGoogle}}, \code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}},\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}},  \code{\link{googleNames}}
}
\keyword{data}
