% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePatentData.R
\name{makeColors}
\alias{makeColors}
\title{Make color hues}
\usage{
makeColors(numColors)
}
\arguments{
\item{numColors}{Number of colors, a numeric input.}
}
\value{
A character vector of colors.
}
\description{
Generate an evenly-spaced number of color hues.
 
Credit for this function goes to \href{http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}{John Colby's}
Stack Overflow post.
}
\examples{
makeColors(5)

}
\seealso{
\code{\link{flippedHistogram}}
}
