% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patent_data.R
\docType{data}
\name{patent_data}
\alias{patent_data}
\title{Patent Dataset}
\format{
A data frame with 51232 rows and 11 variables:
\describe{
\item{\code{guid}}{character. The patent number, a unique identifier.}
\item{\code{inventionTitle}}{character. The title of the patent.}
\item{\code{datePublished}}{POSIXct, date. The date when the patent was published.}
\item{\code{applicationFilingDate}}{character. The date when the patent application was filled.}
\item{\code{inventorsName}}{character. Inventors' names.}
\item{\code{inventorCity}}{character. Inventors' city of residence.}
\item{\code{inventorState}}{character. Inventors' state of residence.}
\item{\code{assigneeName}}{character. Assignees' names.}
\item{\code{assigneeCity}}{character. Assignees' city of residence.}
\item{\code{assigneeState}}{character. Assignees' state of residence.}
\item{\code{cpcInventiveFlattened}}{character. Classifications of the patent.}
}
}
\usage{
patent_data
}
\description{
This dataset contains 51232 patents from the USPTO database.
}
\keyword{datasets}
