% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patentr.R
\docType{package}
\name{patentr}
\alias{patentr}
\alias{patentr-package}
\title{patentr: A package for analyzing patent data.}
\description{
The package is a data processing and reporting tool of patent data sets for patent analysts.
}
\details{
The package is aimed at patent agents, lawyers, managers, analysts, and 
academics who are working on patents. This may be used in a patent landscape 
analysis, company IP portfolio analysis, or a freedom to operate search.

The patentr package provides four categories of important functions:

\enumerate{
\item Data input and cleaning
\item Data (re)processing
\item Data exploration & visualization
\item Visualization & reporting
}
}
\section{patentr functions}{


\code{\link{importPatentData}}: Import xls patent data from filepaths.
}

