% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPatentData.R
\name{readPatentData}
\alias{readPatentData}
\title{Read in a data file or list of files from excel spreadsheets.}
\usage{
readPatentData(rawDataFilePath = NA, skipLines = 1)
}
\arguments{
\item{rawDataFilePath}{A file, or list of files, in xls format. URL in future.}

\item{skipLines}{Number of lines to skip before reading in your data file.}
}
\value{
A single data frame of all data.
}
\description{
This read in file can turn into an S3 class where you read from 
a URL, an excel file, google doc, and a csv.
}
\examples{
add(1, 1)
add(10, 1)

}
