% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{removeDups}
\alias{removeDups}
\title{Remove duplicate values in a patent data set.}
\usage{
removeDups(input)
}
\arguments{
\item{input}{A vector or a data frame which you wish to remove duplicate values.}
}
\value{
A logical vector of TRUE / FALSE values indicating which elements are 
duplciates.
}
\description{
Remove duplicate values in the patent data. Typically you will 
want to check if you have repeat document numbers. A document number should be 
a unique number in your dataset, thus, having a duplicate document number in your 
data set should be avoided. 

Often times, your data sets contain duplicate patent entries. This function is 
a wrapper function of the \code{\link[base]{duplicated}} function, 
applied to a dataframe or vector.
}
\examples{

acars <- acars[removeDups(acars$docNum),]

}
\seealso{
\code{\link[base]{duplicated}}
}
