% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runExample}
\alias{runExample}
\title{Shiny app}
\usage{
runExample()
}
\description{
this is a shiny app that loads patent data, views it,
and does a simple visualization. 

NOTE: This only works with xlsx files.
}
