% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoreColors}
\alias{scoreColors}
\title{Score colors used in graphing.}
\format{A character vector. 

\describe{
\item{scoreColors}{A character variable of four score colors for 0 to 3.}
}}
\usage{
scoreColors
}
\description{
A character vector of Hexadecimal score colors.
}
\seealso{
\code{\link{flippedHistogram}}
}
\keyword{data}
