% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_category.R
\name{separate_category}
\alias{separate_category}
\title{Separate the category column into multiple columns.}
\usage{
separate_category(dataset)
}
\arguments{
\item{dataset}{The dataset to use.}
}
\value{
\code{dataset}: A dataset with the category column separated into multiple columns.
}
\description{
This function serves a special purpose to separate the category \code{cpcInventiveFlattened} column into multiple columns.
}
\examples{
library(patents)
data('patent_data')
x1=separate_category(patent_data)
}
