% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{showDups}
\alias{showDups}
\title{View all your duplicate entries to decide which to remove.}
\usage{
showDups(input)
}
\arguments{
\item{input}{A vector or a data frame which you wish to view duplicated values.}
}
\value{
A logical vector of TRUE / FALSE with all entries that contain two 
or more duplicate values.
}
\description{
Return a logical vector of all duplicate entries. 

Often times, you want to review your duplicate results to determine which 
rows you want to keep and which you want to erase. 

For example, if you have 
an application number that is an application, and another that is a search report, 
then you will want to keep the application data and remove the search report 
entry. 

Or, if you have an application number that has both a grant and an 
application entry, you may want to remove the application from your analysis 
and focus on the grant data, as the claim scope is most important for the 
granted patent.
}
\examples{

acarsDups <- acars[showDups(acars$appNum),]
head(acarsDups[order(acarsDups$appNum),c("docNum","docTypeSumobrain","appNum")])

}
\seealso{
\code{\link[base]{duplicated}}, \code{\link{removeDups}}
}
