% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skipGoogle}
\alias{skipGoogle}
\title{How many lines to skip in a Google Patents CSV export file.}
\format{A numeric value.

\describe{
\item{skipGoogle}{A numeric value for number of lines to skip in a Google 
Patents csv export.}

}}
\usage{
skipGoogle
}
\description{
A hard-coded value for the number of lines to skip in a Google Patents csv 
export.
}
\details{
Used with \code{\link{acarsGoogle}} data.
}
\seealso{
\code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}},\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}}, \code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
