% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skipLens}
\alias{skipLens}
\title{How many lines to skip in a lens.org patent data export.}
\format{A numeric value.

\describe{
\item{skipLens}{A numeric value representing the number of rows to skip in a 
lens.org csv data export.}

}}
\usage{
skipLens
}
\description{
How many lines to skip in a lens.org patent data export.
}
\details{
Used with \code{\link{acarsLens}} data.
}
\seealso{
\code{\link{skipGoogle}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}},\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}}, \code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
