% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skipSumobrain}
\alias{skipSumobrain}
\title{The number of lines to skip in a data read for a sumobrain.com export file.}
\format{A numeric value.

\describe{
\item{skipSumobrain}{A hard-coded numeric value for how many lines to skip 
in a sumobrain.com data export.}

}}
\usage{
skipSumobrain
}
\description{
The number of lines to skip in a data read for a sumobrain.com export file.
Used with \code{\link{acars}} data.
}
\seealso{
\code{\link{skipGoogle}}, \code{\link{skipLens}}, 
\code{\link{googleColumns}},\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{sumobrainNames}}, \code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
