% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorePatentData.R
\name{summarizeColumns}
\alias{summarizeColumns}
\title{Summarize columns of a data frame}
\usage{
summarizeColumns(df, names, naOmit = FALSE)
}
\arguments{
\item{df}{A data frame of patent data.}

\item{names}{a character vector of header names that you want to summarize.}

\item{naOmit}{Logical. Optionally, remove NA values at the end of the summary.
Useful when comparing fields that have NA values, such as features.}
}
\value{
A dataframe of summarize values.
}
\description{
Summarize columns of a data frame.

Summarize a data frame \code{df} by a \code{names} character vector of 
header names.
}
\examples{
sumo <- cleanPatentData(patentData = patentr::acars, columnsExpected = sumobrainColumns,
cleanNames = sumobrainNames,
dateFields = sumobrainDateFields,
dateOrders = sumobrainDateOrder,
deduplicate = TRUE,
cakcDict = patentr::cakcDict,
docLengthTypesDict = patentr::docLengthTypesDict,
keepType = "grant",
firstAssigneeOnly = TRUE, 
assigneeSep = ";",
stopWords = patentr::assigneeStopWords)

# note that in reality, you need a patent analyst to carefully score
# these patents, the score here is for demonstrational purposes
score <- round(rnorm(dim(sumo)[1],mean=1.4,sd=0.9))
score[score>3] <- 3
score[score<0] <- 0
sumo$score <- score
scoreSum <- summarizeColumns(sumo, "score")
scoreSum
# load library(ggplot2) for the below part to run
# ggplot(scoreSum, aes(x=score, y = total, fill=factor(score) )) + geom_bar(stat="identity")
nameAndScore <- summarizeColumns(sumo, c("assigneeClean","score"))
# tail(nameAndScore)

}
