% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPatentData.R
\name{summaryText}
\alias{summaryText}
\title{Add summary text to be used in a pptx slide}
\usage{
summaryText(df, singular, plural, sumVar)
}
\arguments{
\item{df}{A summarized patent data frame, summarized by one variable. 
See \code{\link{summarizeColumns}}.}

\item{singular}{The name of the variable, singular version. A character string.
For example: assignee.}

\item{plural}{The name of the variable, plural version. A character string.
For example: assignees, with an 's'.}

\item{sumVar}{The vector of the variable to summarize, taken from the original
patent data set. For example \code{sumo$score} to summarize the score range.}
}
\value{
A length four character vector.
}
\description{
Add a standard summarized text that will be used in 
association with a plot.
}
\examples{
sumo <- cleanPatentData(patentData = patentr::acars, columnsExpected = sumobrainColumns,
cleanNames = sumobrainNames,
dateFields = sumobrainDateFields,
dateOrders = sumobrainDateOrder,
deduplicate = TRUE,
cakcDict = patentr::cakcDict,
docLengthTypesDict = patentr::docLengthTypesDict,
keepType = "grant",
firstAssigneeOnly = TRUE,
assigneeSep = ";",
stopWords = patentr::assigneeStopWords)

# note that in reality, you need a patent analyst to carefully score
# these patents, the score here is for demonstrational purposes
score <- round(rnorm(dim(sumo)[1],mean=1.4,sd=0.9))
score[score>3] <- 3; score[score<0] <- 0
sumo$score <- score
sumo$assigneeSmall <- strtrim(sumo$assigneeClean,12)
category <- c("system","control algorithm","product","control system", "communication")
c <- round(rnorm(dim(sumo)[1],mean=2.5,sd=1.5))
c[c>5] <- 5; c[c<1] <- 1
sumo$category <- category[c]
feature1 <- c("adaptive", "park", "lane", NA,NA,NA,NA,NA, 
"brake", "steer","accelerate","deactivate")
f <- round(rnorm(dim(sumo)[1],mean=5,sd=1))
l <- length(feature1)
f[f>l] <- l; f[f<1] <- 1
sumo$feature1 <- c(feature1,feature1[f])[1:dim(sumo)[1]]

# Summarize the assignees
as <- summarizeColumns(sumo, 'assigneeSmall')
summaryText(as, 'assignee','assignees',sumo$score)
# summarize the number of features
f <- summarizeColumns(sumo, 'feature1', naOmit = TRUE)
summaryText(f, 'feature','features',sumo$feature1)

}
