% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumobrainColumns}
\alias{sumobrainColumns}
\title{The number of columns in a sumobrain.com data export.}
\format{A numeric value.

\describe{
\item{sumobrainColumns}{A hard-coded numeric value for the number of columns in a 
sumobrain.com data export.}

}}
\usage{
sumobrainColumns
}
\description{
A convenient hard-coded value that can be used when reading in sumobrain.com 
exported patent data files.
}
\details{
Used with \code{\link{acars}} data.
}
\seealso{
\code{\link{skipGoogle}}, \code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}},\code{\link{lensColumns}},
\code{\link{sumobrainNames}}, \code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
