% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumobrainDateFields}
\alias{sumobrainDateFields}
\title{A simple list of date column names in sumobrain data.}
\format{A character vector

\describe{
\item{sumobrainDateFields}{A character vector of date fields.}
}}
\usage{
sumobrainDateFields
}
\description{
A character vector of date fields in sumobrain data.
}
\seealso{
\code{\link{cleanHeaderNames}}
}
\keyword{data}
