% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumobrainDateOrder}
\alias{sumobrainDateOrder}
\title{Date order for sumobrain data.}
\format{A character value.

\describe{
\item{sumobrainDateOrder}{A character variable of date order.}
}}
\usage{
sumobrainDateOrder
}
\description{
A date order to be used in sumobrain date data.
}
\seealso{
\code{\link{extractCleanDate}}
}
\keyword{data}
