% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumobrainNames}
\alias{sumobrainNames}
\title{Header names for a data upload sourced from sumobrain.com}
\format{A character vector

\describe{
\item{sumobrainNames}{An 11-element character vector of clean sumobrain names.}

}}
\usage{
sumobrainNames
}
\description{
A character vector to standardize the headers of the imported excel from a 
sumobrain.com patent data export.
}
\details{
Used with \code{\link{acars}} data.
}
\seealso{
Go to \href{www.sumobrain.com}{Sumobrain}, create a free account, and 
download the data.

\code{\link{skipGoogle}}, \code{\link{skipLens}}, \code{\link{skipSumobrain}},
\code{\link{googleColumns}},\code{\link{lensColumns}}, \code{\link{sumobrainColumns}}, 
\code{\link{lensNames}}, \code{\link{googleNames}}
}
\keyword{data}
