% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePatentData.R
\name{wordCloudIt}
\alias{wordCloudIt}
\title{Generate a word cloude with a given subset of patent data fields.}
\usage{
wordCloudIt(file, rmwords, minfreq = 20, maxwords = 150, ...)
}
\arguments{
\item{file}{The data frame you want word cloud, typically the abstract, title, 
and claims subset.}

\item{rmwords}{A character vector of words you exclude from your analysis. Default 
is \code{\link{excludeWords}}.}

\item{minfreq}{From \code{\link[wordcloud]{wordcloud}}, the min frequency 
to include a word. Default is 10.}

\item{maxwords}{From \code{\link[wordcloud]{wordcloud}}, the max number of 
words to show. Default is 150.}

\item{...}{\code{\link[wordcloud]{wordcloud}} options}
}
\value{
NULL, prints out a wordcloud
}
\description{
Create a word cloud from a patent data set.
}
\examples{

sumo <- cleanPatentData(patentData = patentr::acars, columnsExpected = sumobrainColumns,
cleanNames = sumobrainNames,
dateFields = sumobrainDateFields,
dateOrders = sumobrainDateOrder,
deduplicate = TRUE,
cakcDict = patentr::cakcDict,
docLengthTypesDict = patentr::docLengthTypesDict,
keepType = "grant",
firstAssigneeOnly = TRUE,
assigneeSep = ";",
stopWords = patentr::assigneeStopWords)

# df <- dplyr::select(sumo, title, abstract)
df <- sumo[,c("title","abstract")]
wordCloudIt(df, excludeWords, minfreq = 20, 
random.order = FALSE, rot.per = 0.25)

}
