% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{generateDocType}
\alias{generateDocType}
\title{Calculate the type of document}
\usage{
generateDocType(officeDocLength, countryAndKindCode,
  cakcDict = patentr::cakcDict,
  docLengthTypesDict = patentr::docLengthTypesDict)
}
\arguments{
\item{officeDocLength}{The concat value of country code and number of numerical digits. 
Extracted using the \code{\link{extractDocLength}} function.}

\item{countryAndKindCode}{The concat value of the country code and kind code. 
Extracted using the \code{\link{extractCountryCode}} and \code{\link{extractKindCode}} 
functions.}

\item{cakcDict}{A county and kind code dictionary. Default is \code{\link{cakcDict}}.}

\item{docLengthTypesDict}{A document length and type dictionary. Default is \code{\link{docLengthTypesDict}}.}
}
\value{
A vector of characters labeling the document type, with NA for when 
no match was found.
}
\description{
Determine the type of document from the patent publication data. 

Often times, data exports from publicly available sources do not provide the 
type of patent document, or, if provided, still requires standardization. By 
using the kind code, country code,
}
\examples{

acars <- acars
acars$pubNum <- extractPubNumber(acars$docNum) # pubnum, ex ####
acars$countryCode <- extractCountryCode(acars$docNum) # country code, ex USAPP, USD
acars$officeDocLength <- extractDocLength(countryCode = acars$countryCode, 
                                         pubNum = acars$pubNum) # cc + pub num length concat
acars$kindCode <- extractKindCode(acars$docNum)
acars$countryAndKindCode <- with(acars, paste0(countryCode, kindCode))
                                         
acars$docType <- generateDocType(officeDocLength = acars$officeDocLength,
                            countryAndKindCode = acars$countryAndKindCode,
                            cakcDict = cakcDict,
                            docLengthTypesDict = docLengthTypesDict)
table(acars$docType)


}
\seealso{
\code{\link{cakcDict}}, \code{\link{docLengthTypesDict}}
}
