% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPatentData.R
\name{removeDups}
\alias{removeDups}
\title{Remove duplicate values in a patent data set.}
\usage{
removeDups(input)
}
\arguments{
\item{input}{A vector or a data frame which you wish to remove duplicate values.}
}
\value{
A logical vector of TRUE / FALSE values indicating with one TRUE value 
per duplicate (two or more identical) values.
}
\description{
Remove duplicate values in the patent data. Typically you will 
want to check if you have repeat document numbers. A document number should be 
a unique number in your dataset, thus, having a duplicate document number in your 
data set should be avoided. 

Often times, your data sets contain duplicate patent entries. This function is 
a wrapper function of the \code{\link[base]{duplicated}} function, 
applied to a dataframe or vector. 

For example, if you have the vector [US123, US123, US456], you will get the value 
TRUE FALSE TRUE and the duplicate value is removed.
}
\examples{

acars <- acars[removeDups(acars$docNum),]
head(removeDups(acars$docNum))

}
\seealso{
\code{\link[base]{duplicated}}, \code{\link{showDups}}
}
